@@ -12,9 +12,7 @@ use Symfony\Component\Messenger\HandleTrait;
 use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Routing\Annotation\Route;
 
-/**
- * @Route("/api/categories/", name="api_category_post", methods={"POST"})
- */
+#[Route('/api/categories/', name: 'api_category_post', methods: ['POST'])]
 final class PostCategoryController extends AbstractController
 {
     use HandleTrait;
